avance de herencia

+++ AppSeriesUY/SeriesUY/models.py
@@ -10,14 +10,14 @@ class Usuario(models.Model):
 
 class Serie(models.Model):
     titulo = models.CharField(max_length=30)
-    puntuacion = models.IntegerField()
+    puntuacion = models.FloatField()
     
     def __str__(self):
         return f'Titulo: {self.titulo} Puntuacion: {self.puntuacion}'
     
 class Plataforma(models.Model):
     nombre = models.CharField(max_length=30)
-    precio = models.IntegerField()
+    precio = models.FloatField()
     
     def __str__(self):
         return f'Plataforma: {self.nombre} Precio: {self.precio}'
